The index.html is a minimal website along with the javascript and css to make it functional.
The code uses the same libraries (jquery, bootstrap3 ) as the old ecopad and is written in the
same style where possible. It is however much much shorter and only
intended as an easy to understand starting point.

Some of the original minimal (html( code for the SEV Forecasting has been left in to make it easier to see where the code of the real ecopad would reside.
The javascript code for these parts would not work without adaptation and
has been removed  to make the example as short as possilbe.
One of the next tasks will be to gradually adapt the javascript of the old ecopad to work with the
- new version of cybercommons,
- the new task queue and
- the containers for the models.
as this example.

This first commit also avoids the dependency on bower (a
dependency manager for the javascript, and css libraries running in the browser of an ecopad user.)
It also avoids hanldebar templates  since the version of handlebars uses by ecopad
are outdated and a security risk (github warning)
The original paths of these components have been commented and
temporarily replaced by direct links to the appropriate cdns. (content
delivery networks) for the dependecies that are necessary for the
example.